Move delegate adapter for RecyclerView into separate module.
