--- app/src/main/java/ru/nikijava/androidacademynewsapp/delegate_adapter/Item.java
@@ -1,4 +1,0 @@
-package ru.nikijava.androidacademynewsapp.delegate_adapter;
-
-public interface Item {
-}
