+++ delegateadapter/src/main/java/ru/nikijava/adapterdelegate/BaseViewHolder.java
@@ -1,4 +1,4 @@
-package ru.nikijava.androidacademynewsapp.delegate_adapter;
+package ru.nikijava.adapterdelegate;
 
 import android.support.annotation.NonNull;
 import android.support.v7.widget.RecyclerView;
