+++ delegateadapter/src/main/java/ru/nikijava/adapterdelegate/Item.java
@@ -1,0 +1,4 @@
+package ru.nikijava.adapterdelegate;
+
+public interface Item {
+}
